Перенёс `maxo/utils/urls.py` в `maxo/utils/link.py`
