+++ docs/pages/utils/link.rst
@@ -1,0 +1,7 @@
+Link Utility
+============
+
+.. automodule:: maxo.utils.link
+   :members:
+   :undoc-members:
+   :show-inheritance:
