+++ src/maxo/types/linked_message.py
@@ -4,7 +4,7 @@ from maxo.omit import Omittable, Omitted, is_defined
 from maxo.types.base import MaxoType
 from maxo.types.message_body import MessageBody
 from maxo.types.user import User
-from maxo.utils.urls import id_to_message_url
+from maxo.utils.link import id_to_message_url
 
 
 class LinkedMessage(MaxoType):
@@ -47,6 +47,12 @@ class LinkedMessage(MaxoType):
 
     @property
     def generated_url(self) -> str | None:
+        """
+        Генерирует ссылку на сообщение.
+
+        Возвращает:
+            str | None: Ссылка на сообщение или `None`, если `chat_id` не определен.
+        """
         if not is_defined(self.chat_id):
             return None
 
@@ -57,6 +63,14 @@ class LinkedMessage(MaxoType):
 
     @property
     def unsafe_generated_url(self) -> str:
+        """
+        Генерирует ссылку на сообщение.
+
+        Если `generated_url` не определен, вызывает :py:exc:`~maxo.errors.AttributeIsEmptyError`.
+
+        Возвращает:
+            str: Ссылка на сообщение.
+        """
         if is_defined(self.generated_url):
             return self.generated_url
